requirement.txt
h5py==3.8.0
librosa=0.9.2
numpy==1.23.5
pymp-pypi==0.5.0
pypesq==1.2.4
pystoi==0.3.3
python-gflags==3.1.2
pytorch==1.12.1
pyyaml==6.0
soundfile==0.11.0
thop==0.1.1-2209072238
torchaudio==0.12.1
scipy==1.10.0
